move to bridging utils
