+++ contracts/authorization/src/domain.rs
@@ -1,16 +1,12 @@
-use cosmwasm_std::{
-    to_json_binary, Addr, Binary, CosmosMsg, DepsMut, HexBinary, Storage, Uint64, WasmMsg,
-};
+use cosmwasm_std::{Binary, CosmosMsg, DepsMut, StdResult, Storage, Uint64, WasmMsg};
 use valence_authorization_utils::{
     authorization::{Authorization, Subroutine},
-    domain::{
-        CosmwasmBridge, Domain, EvmBridge, ExecutionEnvironment, ExternalDomain, PolytoneNote,
-    },
+    domain::{CosmwasmBridge, Domain, EvmBridge, ExecutionEnvironment, ExternalDomain},
     msg::ExternalDomainInfo,
 };
 use valence_bridging_utils::{
-    hyperlane::{DispatchMsg, HyperlaneExecuteMsg},
-    polytone::{CallbackRequest, PolytoneExecuteMsg},
+    hyperlane::create_msg_for_hyperlane,
+    polytone::{create_msg_for_polytone, CallbackRequest},
 };
 
 use crate::{
@@ -60,7 +56,7 @@ pub fn create_msg_for_processor(
     execute_msg: Binary,
     domain: &Domain,
     callback_request: Option<CallbackRequest>,
-) -> Result<CosmosMsg, ContractError> {
+) -> StdResult<CosmosMsg> {
     match domain {
         Domain::Main => create_msg_for_main_domain(storage, execute_msg),
         Domain::External(external_domain) => {
@@ -68,7 +64,8 @@ pub fn create_msg_for_processor(
             match external_domain.execution_environment {
                 ExecutionEnvironment::Cosmwasm(cosmwasm_bridge) => match cosmwasm_bridge {
                     CosmwasmBridge::Polytone(polytone_connectors) => create_msg_for_polytone(
-                        polytone_connectors.polytone_note,
+                        polytone_connectors.polytone_note.address.to_string(),
+                        Uint64::from(polytone_connectors.polytone_note.timeout_seconds),
                         external_domain.processor,
                         execute_msg,
                         callback_request,
@@ -90,7 +87,7 @@ pub fn create_msg_for_processor(
 pub fn create_msg_for_main_domain(
     storage: &dyn Storage,
     execute_msg: Binary,
-) -> Result<CosmosMsg, ContractError> {
+) -> StdResult<CosmosMsg> {
     let processor = PROCESSOR_ON_MAIN_DOMAIN.load(storage)?;
     // Simple message for the main domain's processor
     Ok(CosmosMsg::Wasm(WasmMsg::Execute {
@@ -98,54 +95,4 @@ pub fn create_msg_for_main_domain(
         msg: execute_msg,
         funds: vec![],
     }))
-}
-
-pub fn create_msg_for_polytone(
-    polytone_note: PolytoneNote,
-    processor: String,
-    execute_msg: Binary,
-    callback_request: Option<CallbackRequest>,
-) -> Result<CosmosMsg, ContractError> {
-    Ok(CosmosMsg::Wasm(WasmMsg::Execute {
-        contract_addr: polytone_note.address.to_string(),
-        msg: to_json_binary(&PolytoneExecuteMsg::Execute {
-            msgs: vec![CosmosMsg::Wasm(WasmMsg::Execute {
-                contract_addr: processor,
-                msg: execute_msg,
-                funds: vec![],
-            })],
-            callback: callback_request,
-            timeout_seconds: Uint64::from(polytone_note.timeout_seconds),
-        })?,
-        funds: vec![],
-    }))
-}
-
-pub fn create_msg_for_hyperlane(
-    mailbox: Addr,
-    domain_id: u32,
-    processor: String,
-    execute_msg: Binary,
-) -> Result<CosmosMsg, ContractError> {
-    Ok(CosmosMsg::Wasm(WasmMsg::Execute {
-        contract_addr: mailbox.to_string(),
-        msg: to_json_binary(&HyperlaneExecuteMsg::Dispatch(DispatchMsg {
-            dest_domain: domain_id,
-            recipient_addr: format_address_for_hyperlane(processor)?,
-            msg_body: HexBinary::from(execute_msg.to_vec()),
-            hook: None,
-            metadata: None,
-        }))?,
-        funds: vec![],
-    }))
-}
-
-/// Formats an address for Hyperlane by removing the "0x" prefix and padding it to 32 bytes
-pub fn format_address_for_hyperlane(address: String) -> Result<HexBinary, ContractError> {
-    // Remove "0x" prefix if present
-    let address_hex = address.trim_start_matches("0x").to_string().to_lowercase();
-    // Pad to 32 bytes (64 hex characters) because mailboxes expect 32 bytes addresses with leading zeros
-    let padded_address = format!("{:0>64}", address_hex);
-    // Convert to HexBinary which is what Hyperlane expects
-    Ok(HexBinary::from_hex(&padded_address)?)
 }
