+++ packages/bridging-utils/src/polytone.rs
@@ -1,5 +1,7 @@
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::{Addr, Binary, CosmosMsg, Empty, SubMsgResponse, Uint64};
+use cosmwasm_std::{
+    to_json_binary, Addr, Binary, CosmosMsg, Empty, StdResult, SubMsgResponse, Uint64, WasmMsg,
+};
 
 #[cw_serde]
 pub enum PolytoneExecuteMsg {
@@ -49,4 +51,26 @@ pub struct ErrorResponse {
     pub message_index: Uint64,
     /// The error that occured executing the message.
     pub error: String,
+}
+
+pub fn create_msg_for_polytone(
+    polytone_note: String,
+    timeout_seconds: Uint64,
+    processor: String,
+    execute_msg: Binary,
+    callback_request: Option<CallbackRequest>,
+) -> StdResult<CosmosMsg> {
+    Ok(CosmosMsg::Wasm(WasmMsg::Execute {
+        contract_addr: polytone_note,
+        msg: to_json_binary(&PolytoneExecuteMsg::Execute {
+            msgs: vec![CosmosMsg::Wasm(WasmMsg::Execute {
+                contract_addr: processor,
+                msg: execute_msg,
+                funds: vec![],
+            })],
+            callback: callback_request,
+            timeout_seconds,
+        })?,
+        funds: vec![],
+    }))
 }
